# contrato/openapi.json
{
  "components": {
    "schemas": {
      "Fuente": {
        "properties": {
          "post_id": {
            "title": "Post Id",
            "type": "integer"
          },
          "puntuacion": {
            "title": "Puntuacion",
            "type": "number"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "titulo": {
            "title": "Titulo",
            "type": "string"
          }
        },
        "required": [
          "post_id",
          "slug",
          "titulo",
          "puntuacion"
        ],
        "title": "Fuente",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MotivoOmision": {
        "properties": {
          "motivo": {
            "title": "Motivo",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "motivo"
        ],
        "title": "MotivoOmision",
        "type": "object"
      },
      "PeticionBuscar": {
        "properties": {
          "consulta": {
            "maxLength": 500,
            "minLength": 2,
            "title": "Consulta",
            "type": "string"
          },
          "limite": {
            "default": 5,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Limite",
            "type": "integer"
          }
        },
        "required": [
          "consulta"
        ],
        "title": "PeticionBuscar",
        "type": "object"
      },
      "PeticionIndexar": {
        "properties": {
          "posts": {
            "items": {
              "$ref": "#/components/schemas/PostEntrante"
            },
            "title": "Posts",
            "type": "array"
          }
        },
        "required": [
          "posts"
        ],
        "title": "PeticionIndexar",
        "type": "object"
      },
      "PeticionPreguntar": {
        "properties": {
          "consulta": {
            "maxLength": 500,
            "minLength": 2,
            "title": "Consulta",
            "type": "string"
          },
          "limite": {
            "default": 3,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Limite",
            "type": "integer"
          }
        },
        "required": [
          "consulta"
        ],
        "title": "PeticionPreguntar",
        "type": "object"
      },
      "PeticionResumir": {
        "properties": {
          "texto": {
            "maxLength": 20000,
            "minLength": 20,
            "title": "Texto",
            "type": "string"
          }
        },
        "required": [
          "texto"
        ],
        "title": "PeticionResumir",
        "type": "object"
      },
      "PostEntrante": {
        "description": "Un post tal y como lo manda blog-api para indexarlo.",
        "properties": {
          "categoria": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoria"
          },
          "cuerpo": {
            "title": "Cuerpo",
            "type": "string"
          },
          "estado": {
            "enum": [
              "borrador",
              "publicado"
            ],
            "title": "Estado",
            "type": "string"
          },
          "etiquetas": {
            "items": {
              "type": "string"
            },
            "title": "Etiquetas",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "publicado_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publicado En"
          },
          "resumen": {
            "title": "Resumen",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "titulo": {
            "title": "Titulo",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "titulo",
          "resumen",
          "cuerpo",
          "estado"
        ],
        "title": "PostEntrante",
        "type": "object"
      },
      "RespuestaBuscar": {
        "properties": {
          "consulta": {
            "title": "Consulta",
            "type": "string"
          },
          "resultados": {
            "items": {
              "$ref": "#/components/schemas/ResultadoBusqueda"
            },
            "title": "Resultados",
            "type": "array"
          }
        },
        "required": [
          "consulta",
          "resultados"
        ],
        "title": "RespuestaBuscar",
        "type": "object"
      },
      "RespuestaIndexar": {
        "properties": {
          "fragmentos": {
            "title": "Fragmentos",
            "type": "integer"
          },
          "indexados": {
            "title": "Indexados",
            "type": "integer"
          },
          "motivos_omision": {
            "items": {
              "$ref": "#/components/schemas/MotivoOmision"
            },
            "title": "Motivos Omision",
            "type": "array"
          },
          "omitidos": {
            "title": "Omitidos",
            "type": "integer"
          }
        },
        "required": [
          "indexados",
          "omitidos",
          "fragmentos",
          "motivos_omision"
        ],
        "title": "RespuestaIndexar",
        "type": "object"
      },
      "RespuestaPreguntar": {
        "properties": {
          "consulta": {
            "title": "Consulta",
            "type": "string"
          },
          "fuentes": {
            "items": {
              "$ref": "#/components/schemas/Fuente"
            },
            "title": "Fuentes",
            "type": "array"
          },
          "modelo": {
            "title": "Modelo",
            "type": "string"
          },
          "respuesta": {
            "title": "Respuesta",
            "type": "string"
          }
        },
        "required": [
          "consulta",
          "respuesta",
          "fuentes",
          "modelo"
        ],
        "title": "RespuestaPreguntar",
        "type": "object"
      },
      "RespuestaResumir": {
        "properties": {
          "modelo": {
            "title": "Modelo",
            "type": "string"
          },
          "resumen": {
            "title": "Resumen",
            "type": "string"
          }
        },
        "required": [
          "resumen",
          "modelo"
        ],
        "title": "RespuestaResumir",
        "type": "object"
      },
      "RespuestaSalud": {
        "properties": {
          "estado": {
            "title": "Estado",
            "type": "string"
          },
          "fragmentos_indexados": {
            "title": "Fragmentos Indexados",
            "type": "integer"
          },
          "modelo_embeddings": {
            "title": "Modelo Embeddings",
            "type": "string"
          },
          "modelo_generacion": {
            "title": "Modelo Generacion",
            "type": "string"
          },
          "ollama_alcanzable": {
            "title": "Ollama Alcanzable",
            "type": "boolean"
          },
          "posts_indexados": {
            "title": "Posts Indexados",
            "type": "integer"
          },
          "servicio": {
            "title": "Servicio",
            "type": "string"
          }
        },
        "required": [
          "estado",
          "servicio",
          "modelo_embeddings",
          "modelo_generacion",
          "ollama_alcanzable",
          "posts_indexados",
          "fragmentos_indexados"
        ],
        "title": "RespuestaSalud",
        "type": "object"
      },
      "ResultadoBusqueda": {
        "properties": {
          "fragmento": {
            "title": "Fragmento",
            "type": "string"
          },
          "post_id": {
            "title": "Post Id",
            "type": "integer"
          },
          "puntuacion": {
            "title": "Puntuacion",
            "type": "number"
          },
          "resumen": {
            "title": "Resumen",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "titulo": {
            "title": "Titulo",
            "type": "string"
          }
        },
        "required": [
          "post_id",
          "slug",
          "titulo",
          "resumen",
          "puntuacion",
          "fragmento"
        ],
        "title": "ResultadoBusqueda",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Busqueda semantica y RAG para el blog «Corriente». Lo consume blog-api; el navegador nunca habla directamente con este servicio.",
    "title": "blog-ai",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/buscar": {
      "post": {
        "operationId": "buscar_buscar_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeticionBuscar"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaBuscar"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Busqueda semantica"
      }
    },
    "/indexar": {
      "post": {
        "description": "Los posts en borrador se omiten y se retiran del indice si ya estaban.",
        "operationId": "indexar_indexar_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeticionIndexar"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaIndexar"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Indexa posts publicados"
      }
    },
    "/preguntar": {
      "post": {
        "operationId": "preguntar_preguntar_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeticionPreguntar"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaPreguntar"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pregunta al blog (RAG)"
      }
    },
    "/resumir": {
      "post": {
        "operationId": "resumir_resumir_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeticionResumir"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaResumir"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Limite de llamadas por cliente superado"
          }
        },
        "summary": "Resume un texto en dos frases"
      }
    },
    "/salud": {
      "get": {
        "operationId": "salud_salud_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaSalud"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Estado del servicio"
      }
    }
  }
}